# openapi.yaml
openapi: 3.0.3
info:
  title: Hi-ring ATS/CRM API
  description: |
    Recruitment platform API providing ATS (Applicant Tracking System) and CRM capabilities.

    ## Authentication
    All endpoints require authentication via NextAuth.js session cookies.

    ## Rate Limiting
    - General endpoints: 10 requests/minute
    - Authentication endpoints: 5 requests/minute

    ## Pagination
    List endpoints support pagination with `page` and `limit` query parameters.
    Default limit: 20, Max limit: 100

    ## Status Codes
    - 200: Success
    - 400: Bad Request (validation error)
    - 401: Unauthorized (no session)
    - 403: Forbidden (insufficient permissions)
    - 404: Not Found
    - 409: Conflict (duplicate resource)
    - 500: Internal Server Error

  version: 1.0.0
  contact:
    name: Hi-ring Support
    email: support@hi-ring.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://staging.hi-ring.com
    description: Staging server
  - url: https://hi-ring.com
    description: Production server
...
paths:
  /api/candidates:
    get:
      tags:
        - Candidates
      summary: List and search candidates
      description: |
        Retrieve a paginated list of candidates with optional filtering and sorting.
        Supports text search, status filtering, skills filtering, and more.
      operationId: getCandidates
      parameters:
        - name: page
          in: query
          description: Page number (1-indexed)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Results per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          description: Text search across name, email, position, company
          schema:
            type: string
        - name: status
          in: query
          description: Filter by status (comma-separated)
          schema:
            type: string
            example: "new,contacted,screening"
        - name: experienceLevel
          in: query
          description: Filter by experience level (comma-separated)
          schema:
            type: string
            example: "junior,mid,senior"
        - name: skills
          in: query
          description: Filter by skills (comma-separated)
          schema:
            type: string
            example: "javascript,react,node.js"
        - name: assignedTo
          in: query
          description: Filter by assigned user ID
          schema:
            type: string
        - name: sortBy
          in: query
          description: Field to sort by
          schema:
            type: string
            enum: [createdAt, updatedAt, lastContactedAt, rating, name]
            default: createdAt
        - name: sortOrder
          in: query
          description: Sort direction
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  candidates:
                    type: array
                    items:
                      $ref: '#/components/schemas/CandidateSummary'
                  total:
                    type: integer
                    description: Total number of candidates matching filters
                  page:
                    type: integer
                  limit:
                    type: integer
                  totalPages:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - sessionCookie: []

    post:
      tags:
        - Candidates
      summary: Create a new candidate
      description: Create a new candidate in the system
      operationId: createCandidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCandidateRequest'
      responses:
        '200':
          description: Candidate created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Candidate created successfully"
                  candidateId:
                    type: string
                    example: "507f1f77bcf86cd799439011"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Conflict - Email already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Candidate with this email already exists"
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - sessionCookie: []
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie
  schemas:
    CandidateSummary:
      type: object
      properties:
        id:
          type: string
          example: "507f1f77bcf86cd799439011"
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        phone:
          type: string
          example: "+33612345678"
        currentPosition:
          type: string
          example: "Senior Developer"
        currentCompany:
          type: string
          example: "Tech Corp"
        experienceLevel:
          $ref: '#/components/schemas/ExperienceLevel'
        status:
          $ref: '#/components/schemas/CandidateStatus'
        skills:
          type: array
          items:
            $ref: '#/components/schemas/Skill'
        overallRating:
          type: number
          minimum: 1
          maximum: 5
          nullable: true
        assignedTo:
          type: string
          nullable: true
        assignedToName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateCandidateRequest:
      type: object
      required:
        - firstName
        - lastName
        - email
        - experienceLevel
        - source
      properties:
        firstName:
          type: string
          minLength: 1
          example: "Jane"
        lastName:
          type: string
          minLength: 1
          example: "Smith"
        email:
          type: string
          format: email
          example: "jane.smith@example.com"
        phone:
          type: string
          example: "+33612345678"
        currentPosition:
          type: string
          example: "Frontend Developer"
        experienceLevel:
          $ref: '#/components/schemas/ExperienceLevel'
        skills:
          type: array
          items:
            $ref: '#/components/schemas/Skill'
        source:
          type: string
          example: "LinkedIn"
        availability:
          $ref: '#/components/schemas/AvailabilityStatus'
        contractPreference:
          type: array
          items:
            $ref: '#/components/schemas/ContractPreference'
        status:
          $ref: '#/components/schemas/CandidateStatus'
        assignedTo:
          type: string
          description: User ID to assign candidate to
...
    Skill:
      type: object
      required:
        - name
        - level
      properties:
        name:
          type: string
          example: "JavaScript"
        level:
          type: string
          enum: [beginner, intermediate, advanced, expert]
          example: "expert"
        yearsOfExperience:
          type: number
          example: 5
        lastUsed:
          type: string
          example: "2024"
        certified:
          type: boolean
...
    CandidateStatus:
      type: string
      enum:
        - new
        - contacted
        - screening
        - interview_scheduled
        - interview_completed
        - offer_sent
        - offer_accepted
        - offer_rejected
        - hired
        - rejected
        - on_hold
        - archived
    ExperienceLevel:
      type: string
      enum:
        - entry
        - junior
        - mid
        - senior
        - lead
        - executive
    AvailabilityStatus:
      type: string
      enum:
        - immediate
        - two_weeks
        - one_month
        - two_months
        - three_months
        - negotiable
        - not_available
    ContractPreference:
      type: string
      enum:
        - cdi
        - cdd
        - freelance
        - stage
        - alternance
        - interim
        - any
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        details:
          description: Additional error details
          oneOf:
            - type: string
            - type: object
            - type: array
              items:
                type: object
  responses:
    Unauthorized:
      description: Unauthorized - No valid session
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Unauthorized"
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Forbidden: Insufficient permissions"
    BadRequest:
      description: Bad Request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid input"
            details:
              - field: "email"
                message: "Invalid email format"
...
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"